Colocando a regra de negocio da HomeViewController na HomeViewModel e criando uma extensio de delegate da tableView
